Split Sell into ItemChoice + SellChosen, update ItemChoice
Split sell into a sequence of choosing what you want to sell, then
selling it. This relies on the following changes:

ItemChoice used to be the one place in the codebase that uses "choices"
to be "the things that I have chosen", as opposed to "the things I may
choose from". Fix this to be more consistent

Add in a nothing_option to item_choice, which gives the possibility for
the player to choose no items.